Move RollingType to mod roller, add trait Roller
