+++ Cargo.toml
@@ -36,3 +36,4 @@ zip = { version = "0.6", optional = true }
 lz4_flex = { version = "0.11", optional = true }
 flate2 = { version = "1.0", optional = true}
 memmap = { version ="0.7.0", optional = true}
+dyn-clone = "1.0.13"
