+++ src/config.rs
@@ -4,7 +4,8 @@ use crate::filter::{Filter, NoFilter};
 use crate::plugin::console::ConsoleAppender;
 use crate::plugin::file::FileAppender;
 use crate::plugin::file_loop::FileLoopAppender;
-use crate::plugin::file_split::{FileSplitAppender, Packer, RawFile, RollingType, SplitFile};
+use crate::plugin::file_split::{FileSplitAppender, Packer, RawFile, SplitFile};
+use crate::plugin::roller::RollingType;
 use crate::FastLogFormat;
 use dark_std::sync::SyncVec;
 use log::LevelFilter;
@@ -98,8 +99,16 @@ impl Config {
         packer: P,
     ) -> Self {
         self.appends.push(Mutex::new(Box::new(
-            FileSplitAppender::<RawFile>::new(file_path, temp_size, rolling_type, Box::new(packer))
-                .unwrap(),
+            FileSplitAppender::<RawFile>::new(
+                file_path,
+                temp_size,
+                Box::new(rolling_type),
+                Box::new(packer),
+            )
+            .unwrap(),
+        )));
+        self
+    }
         )));
         self
     }
@@ -126,8 +135,13 @@ impl Config {
         packer: P,
     ) -> Self {
         self.appends.push(Mutex::new(Box::new(
-            FileSplitAppender::<S>::new(file_path, temp_size, rolling_type, Box::new(packer))
-                .unwrap(),
+            FileSplitAppender::<S>::new(
+                file_path,
+                temp_size,
+                Box::new(rolling_type),
+                Box::new(packer),
+            )
+            .unwrap(),
         )));
         self
     }
