+++ src/plugin/mod.rs
@@ -5,3 +5,4 @@ pub mod file_loop;
 pub mod file_mmap;
 pub mod file_split;
 pub mod packer;
+pub mod roller;
