+++ tests/split_test.rs
@@ -4,6 +4,7 @@ mod test {
     use fast_log::consts::LogSize;
     use fast_log::plugin::file_split::{FileSplitAppender, RawFile, RollingType};
     use fast_log::plugin::packer::LogPacker;
+    use fast_log::plugin::roller::Roller;
     use log::Level;
     use std::fs::remove_dir_all;
     use std::thread::sleep;
@@ -15,7 +16,7 @@ mod test {
         let appender = FileSplitAppender::<RawFile>::new(
             "target/test/",
             LogSize::MB(1),
-            RollingType::All,
+            Box::new(RollingType::All),
             Box::new(LogPacker {}),
         )
         .unwrap();
@@ -32,7 +33,8 @@ mod test {
         }]);
         appender.send_pack();
         sleep(Duration::from_secs(1));
-        let rolling_num = RollingType::KeepNum(0).do_rolling("temp.log", "target/test/");
+        let rolling_num = (Box::new(RollingType::KeepNum(0)) as Box<dyn Roller>)
+            .do_rolling("temp.log", "target/test/");
         assert_eq!(rolling_num, 1);
         let _ = remove_dir_all("target/test/");
     }
